fix(backend): fixed issues in session ender

+++ backend/cmd/ender/main.go
@@ -68,12 +68,12 @@ func main() {
 			consumer.Close()
 			os.Exit(0)
 		case <-tick:
-			failedSessionEnds := make(map[uint64]int64)
+			failedSessionEnds := make(map[uint64]uint64)
 			duplicatedSessionEnds := make(map[uint64]uint64)
 
 			// Find ended sessions and send notification to other services
-			sessions.HandleEndedSessions(func(sessionID uint64, timestamp int64) bool {
-				msg := &messages.SessionEnd{Timestamp: uint64(timestamp)}
+			sessions.HandleEndedSessions(func(sessionID uint64, timestamp uint64) bool {
+				msg := &messages.SessionEnd{Timestamp: timestamp}
 				currDuration, err := pg.GetSessionDuration(sessionID)
 				if err != nil {
 					log.Printf("getSessionDuration failed, sessID: %d, err: %s", sessionID, err)
